Adding the changes of the anims before I leave
Added the commit for the changes for the nuke powerup animation it still
is a little unfinished as it can be tweaked to look better but these are
just minor adjustments like the speed of the anim or scaling it to
match the hitbox/powerup scale. Ive added the files for the slowness and
original explosion powerup animation sprite sheet as I wasnt able to add
the other effect in time so this can either be done by someone else,
left till I come back or scrapped entirely. You got me on
whatsapp/snapchat and stuff though if you need to question about the
animation sprite sheets or anything. The slowness can just be added the
same way but just using the green explosion sprite sheet and attaching
it to the slowness powerup.