Update to hit-combo - fully functional
The hit-combo (var streak in the script) now works as initially
intended. Destroying 5 consecutive asteroids (inclduing power ups)
increases the score multiplier by 1. When player misses or if an
asteroid hits the earth the streak is reset.

What still needs to be done:
There is no visual feedback of what the streak (or multiplier) currently
is.