Added new shield powerup (purple shooting star)
Added a new power up that will have the effect of the planets texture
changing to a metal layer being a shield which will last for a certain
amount of time and protect the planet from damage. the shooting star has
been implemented with sounds and textures and it dissapears on click and for
the metal layer functionality needs to be added which Lenny will take care of.