Many changes - More detail in the lower part
Fixed the scroll bar - Managed to fix the scroll bar by resizing the Rect on the List Background without the child objects attached and setting it to clamped mode managed to work.

Improve the look of the particle effect when the earth gets hit by an asteroid shoots faster and is smaller

More skin unlocks - Added 6 new skins for variations these are sports themed.(Some of the skins look a little sketchy but can be fixed easy with improved textures was done mostly for a test)

Adjusted skin prices - Changed the price of skins around with the new unlocks and changed them to match the amount earned per game

Added in the settings menu with a basic looking design that will be improved upon the scene has a working quality choice however this resets on start or scene change so we will need to implement a PlayerPrefs the sound options are currently not hooked up but will be later

Added in the leaderboard scene don't know if we will even use this but at least all the buttons on the main menu now work

Polished the pause menu a bit to give it a nice border and make all the buttons in the center blocking the planet cause all the buttons was spread away from each other

Adjusted the pause icon look and colours to resemble a pause button and changed the mute symbol so that it has a red glow for when muted for a bit of extra visual information

Adjusted the particle effect for the explosion of the mine to give it more of an effect that stands out so I increased the speed an amount slightly to a nice point

Added 3 particle effects in a ParticleSystem Prefabs folder that are set to be linked to the Power ups they each are labeled and have similar effects just with different explosions these will be set up so when the particle is clicked it will play the effect.

Added icons that appear when a power up is active so you can see a Shield/Slowness Icon when that specific power up is active

Further work on adding a particle explosion for the earth at the end of the game the particle effect isn't fully completed as smoke and added effect can be added all that is left to get a fitting sound clip and adding each to the code and making it work together with a pause for the score final calculator